Updated function names for SIE EP, added defines.

+++ platform/nxp/lpc17xx/usb_base.c
@@ -5,7 +5,6 @@
  */
 
 #include <assert.h>
-#include <stdlib.h>
 #include <xcore/memory.h>
 #include <halm/platform/nxp/lpc17xx/system.h>
 #include <halm/platform/nxp/lpc17xx/usb_base.h>
@@ -67,7 +66,7 @@ static void configPins(struct UsbBase *device,
       config->dm, config->dp, config->connect, config->vbus
   };
 
-  for (unsigned int index = 0; index < ARRAY_SIZE(pinArray); ++index)
+  for (size_t index = 0; index < ARRAY_SIZE(pinArray); ++index)
   {
     if (pinArray[index])
     {
